+++ noodles-util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * util/variant/fs: Add `open` function.
+
 ## 0.69.0 - 2025-07-12
 
 ### Added